pyproject.toml
[project]
name = "graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.10.3",
    "pyzmq>=26.2.0",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.24.0",
    "pytest>=8.3.4",
]

[tool.pytest.ini_options]
pythonpath = ["./src"]
filterwarnings = "ignore::DeprecationWarning"
asyncio_default_fixture_loop_scope = "function"  